Fixes a problem where the signed URL didn't match the URL that was actually invoked
The sign_get_request function signs a URL query string made up of different parts that are
generated into three strings - saml_request_data, sigalg_data, and relay_state_data.

These strings are combined and signed by the below:

```
sign.update(saml_request_data + relay_state_data + sigalg_data)
```

saml_request_data contains _either_ a SAMLRequest or a SAMLResponse.
relay_state_data contains the relay state, or ''
sigalg contains a static string that denotes the signing algorithm

However, the query object that is generated and returned by sign_get_request
*always* contains all of the following arguments, even if they are null/blank/undefined:

SAMLRequest
SAMLResponse
RelayState
SigAlg
Signature

This causes any SAML IdP that checks the signature to invalidate the request, because, for example,
the signed query string for an intial SAMLRequest would be something like

```
?SAMLResponse=&SAMLRequest=<some saml request>&RelayState=&SigAlg=<The signing algorithm>&Signature=<The signature>
```

... while the query string that was signed is more like
```
?SAMLRequest=<some saml request>&SigAlg=<The signing algorithm>
```

This causes the request to fail.  The change in this commit only adds the arguments to the query string that are
needed for the particular request.