Port multiple signature support from saml2tu
As the saml2tu repo isn't a true fork (it has a single commit of all the saml2
history) there's no simple way to merge it in. I've manually ported the
relevant code change it introduced, including the test for it.

The behavioral change this introduces to the saml2 repo:

Previous behavior:
If there are multiple signatures in a SAML document, it will be rejected.

New behavior:
If there are multiple signatures, the first will be checked and used if valid.
Other signatures beyond the first will be ignored.