Copy all namespace declarations from Response
The `add_namespaces_to_child_assertions` function [assumes the presence of an
InclusiveNamespaces
element](https://github.com/Clever/saml2/blob/6cad30fb1e0e9a59ae850f7de6bed4310a403ba9/lib/saml2.coffee#L399)
from which it will read a list of namespaces.

If this is present, any namespace definitions declared by InclusiveNamespaces
are copied from the Response element to the Assertion. Then, when the Assertion
element is handled in isolation from the rest of the Response document, the
namespaces are all still defined.

However if the InclusiveNamespaces element is not found, no namespaces are
copied. If then the Assertion element is namespaced (e.g. `<saml:Assertion>`),
it [will not be
found](https://github.com/Clever/saml2/blob/6cad30fb1e0e9a59ae850f7de6bed4310a403ba9/lib/saml2.coffee#L440)
and the document will be treated as not having an assertion at all.

If there is an InclusiveNamespaces element, we will continue to use that as
before to read the namespaces to apply to the Assertion element.

However if there is none, we'll just read all the namespaces from the Response
element onto the Assertion element.

There seemed no value in asserting the presence of a Signature element within
the Assertion, as it's valid to just have the whole message signed, so I
removed that.