Updated debug to fix a security vulnerability
`debug`, uses `ms`, and your version of `debug` had an older version of `ms` thus updating `debug` fixes the vulnerability found in `ms`. In addition to that I set the module version to be the exact version so when in production you always know which version you get. Hope this helps :)