Validate NotBefore and NotOnOrAfter conditions
SAML responses can (and normally do) contain NotBefore and NotOnOrAfter conditions, which specify when the assertion should be considered valid.

See section 2.5.1.2 of http://docs.oasis-open.org/security/saml/v2.0/saml-core-2.0-os.pdf

These are now checked and if the current time is outside this window the response is rejected with an error message of 'SAML Response is not yet valid' or 'SAML Response is no longer valid'.

post_assert can now accept an option `ignore_timing` which, when `true`, skips over these checks. It's not recommended for use in production, so doesn't need to be documented, but I've used it in the existing tests to avoid changing the dates in all the existing fixtures.