Validate AudienceRestriction condition
If `options.audience` is set, and the SAML response contains an `<AudienceRestriction>` condition, at least one of the `<Audience>` nodes' value must match the `options.audience`.

The `audience` option can be either a string or a regular expression, and defaults to `entity_id` if not set. It can be passed as an option to the `ServiceProvider` constructor, or in a `post_assert()` call.