fix: Remove prepublish script
This script actually did nothing; make build doesn't do anything. So, we'll
remove it so npm does not alert us on "prepublish" being deprecated.

❯ npm i
npm WARN prepublish-on-install As of npm@5, `prepublish` scripts are deprecated.
npm WARN prepublish-on-install Use `prepare` for build steps and `prepublishOnly` for upload-only.
npm WARN prepublish-on-install See the deprecation note in `npm help scripts` for more information.

> saml2-js@2.0.7 prepublish /[...]/Clever/saml2
> make build

make: Nothing to be done for `build'.