dep: Bump async from ^2.5.0 to ^3.2.0
This addresses async being an outdated dependency.

https://github.com/caolan/async/blob/b7361922b1fd231cefa13ff80bfd359f482b5e6e/CHANGELOG.md#v300
notes the major difference between v2 and v3:

  "Most Async methods return a Promise when the final callback is omitted,
   making them await-able!"

In our usage of async in the library, we explicitly have callback functions
specified.

In our tests, we seem to do so as well. Tests (and coverage) runs fine, so I'm
fine with this dependency bump.