dep: Bump xmldom from ^0.1.0 to ^0.4.0
This one is tricky. We primarily use it for DOMParser(), but we also use it for
XMLSerializer(). Any major changes to these functions should give us pause.

https://github.com/xmldom/xmldom/compare/v0.1.31...0.2.1

The main changes here stem from HTML entity parsing and some CI changes. Should
be fine to bump up.

0.3.0 added a new function (getElementsByClassName()) as well as some
restructuring of the library, but that's also fine.

https://github.com/xmldom/xmldom/compare/0.3.0...0.4.0

0.4.0 strictly checks that the type for parseFromString() is a string. In all
cases in the library where we use this, we make sure the input to the function
is a string.