dep: Bump xmlbuilder ~2.2.0 to xmlbuilder2 ^2.4.0
This requires a small code change for tests to still pass.

On xmlbuilder ~2.2.0, XML creation worked with an array of objects. What this
means in practice is that we were able to use a block like:

```javascript
[
  { 'saml:AuthnContextClassRef': 'context:class' },
  { '@Comparison': 'exact' }
]
```

to make an XML block that looked like:

```xml
    <RequestedAuthnContext Comparison="exact">
        <saml:AuthnContextClassRef>context:class</saml:AuthnContextClassRef>
    </RequestedAuthnContext>
```

Any xmlbuilder version past that would generate XML that looked like:

```xml
    <RequestedAuthnContext>
        <saml:AuthnContextClassRef>context:class</saml:AuthnContextClassRef>
    </RequestedAuthnContext>
    <RequestedAuthnContext Comparison="exact"/>
```

However, if we change the context node to be a flat object:

```javascript
  {
    'saml:AuthnContextClassRef': ['context:class', 'context:two'],
    '@Comparison': 'exact'
  }
```

we pass the tests.