test: Add test case to correct for additional SPSSODescriptors
When we upgraded to xmlbuilder2, the method we were using to create XML
documents slightly changed. See 57b1486fffa651344fdf1bcecf5750cada0b42d4 and
c910edbb63c6057d8680fa96cddbd8dd8a75f93f for more context.

This change explicitly checks for the amount of nodes that refer to
SPSSODescriptors. Before we upgraded xmlbuilder, there should only be one. When
we upgraded to xmlbuilder2, it registered 5. With this change to
create_metadata, it is now 1 again.