ci: Update pipeline to support major versions sanely
cimg/node only supports major.minor releases. The next time someone
tests or runs this library, we should only have to be concerned that
the major versions are proper.

Following [1], we can simply have a sane base image, and use an orb
to augment the node capabilities we need to use.

Orb is fixed to a known good patch version.

[1] https://github.com/CircleCI-Public/cimg-node/issues/130