Formatted both saml2.coffee files - one that is the library and the one that deals with the tests.
Added 2 new options for the post and redirect assert
  - notonorafter_required - defaults to false (current behavior)
  - notbefore_required    - defaults to false (current behavior)

Added 4 new test xml files to test the new options that check for the two tags in the Conditions element.
One of the tests checks if the Conditions tag is defined, but one or more of the two new options are set to true - in that case the Conditions tag is required by default.